backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pandas==2.2.2
zhipuai>=2.1.4
python-dotenv==1.0.1
aiofiles==24.1.0
sqlalchemy==2.0.35
